making tests work, updating webrtc version.
The tests were failing with an exception claiming that wrtc needs to be
run in headless mode, which made little sense.

Looking at the code, it seemed that the tests were not passing in a
docset into the Network.  weird.

However, the store was creating a Network as well (and not passing in
wrtc, though it was passing in a docset).

I got the tests to run, and pass, my sending in a function to create a
Network given a docset.   This may not be the best idea, but I am not
even sure if I am going to use this library...